complete 4b

+++ part4/bloglist-backend/controllers/blogs.js
@@ -18,7 +18,7 @@ blogRouter.get('/:id', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
     const blog = new Blog(request.body)
     const savedBlog = await blog.save()
-    if (savedBlog){
+    if (savedBlog) {
         response.json(savedBlog.toJSON())
     } else {
         response.status(400) // bad request
@@ -26,8 +26,31 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(404).end()
+    }
+})
+
+blogRouter.put('/:id', async (request, response) => {
+    const body = request.body
+    const newBlogContent = {
+        author: body.author,
+        title: body.title,
+        url: body.url,
+        likes: body.likes,
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+        await Blog.findByIdAndUpdate(request.params.id, newBlogContent, { new: true })
+        response.status(204).end()
+    } else {
+        response.status(404).end()
+    }
+
 })
 
 module.exports = blogRouter
